Develop

+++ services/coupon/src/eShop.CouponService.Domain/Models/Coupon.cs
@@ -5,6 +5,6 @@ namespace eShop.CouponService.Domain;
 public class Coupon : Entity<Guid> 
 {
     public string CouponCode { get; set; }
-    public decimal DiscountAmount { get; set; }
+    public double DiscountAmount { get; set; }
     public int MinAmount { get; set; }
 }
